feat(content_templates): rename contents_templates into content_templates
