+++ lib/v0/content_templates/base.js
