+++ lib/v0/content_templates/create.js
@@ -2,7 +2,7 @@ const base = require('./base')
 const defaultResponse = require('./default-response')
 
 module.exports.request = {
-  $id: 'https://schemas.tillhub.com/v0/contents_templates.create.request.schema.json',
+  $id: 'https://schemas.tillhub.com/v0/content_templates.create.request.schema.json',
   $schema: 'http://json-schema.org/draft-07/schema#',
   additionalProperties: false,
   type: 'object',
@@ -13,7 +13,7 @@ module.exports.request = {
 }
 
 module.exports.response = {
-  $id: 'https://schemas.tillhub.com/v0/contents_templates.create.response.schema.json',
+  $id: 'https://schemas.tillhub.com/v0/content_templates.create.response.schema.json',
   $schema: 'http://json-schema.org/draft-07/schema#',
   additionalProperties: true,
   type: 'object',
