+++ lib/v0/content_templates/default-response.js
