+++ lib/v0/content_templates/index.js
