+++ lib/v0/content_templates/update.js
@@ -2,7 +2,7 @@ const base = require('./base')
 const defaultResponse = require('./default-response')
 
 module.exports.request = {
-  $id: 'https://schemas.tillhub.com/v0/contents_templates.update.request.schema.json',
+  $id: 'https://schemas.tillhub.com/v0/content_templates.update.request.schema.json',
   $schema: 'http://json-schema.org/draft-07/schema#',
   additionalProperties: false,
   type: 'object',
@@ -11,7 +11,7 @@ module.exports.request = {
 }
 
 module.exports.response = {
-  $id: 'https://schemas.tillhub.com/v0/contents_templates.update.response.schema.json',
+  $id: 'https://schemas.tillhub.com/v0/content_templates.update.response.schema.json',
   $schema: 'http://json-schema.org/draft-07/schema#',
   additionalProperties: true,
   type: 'object',
