+++ lib/v0/index.js
@@ -20,7 +20,7 @@ module.exports.favourites = require('./favourites')
 module.exports.functions = require('./functions')
 module.exports.devices = require('./devices')
 module.exports.contents = require('./contents')
-module.exports.contents_templates = require('./contents_templates')
+module.exports.content_templates = require('./content_templates')
 module.exports.reasons = require('./reasons')
 module.exports.tags = require('./tags')
 module.exports.safes = require('./safes')
